+++ src/domain/application/use-cases/task/delete-task.ts
@@ -1,9 +1,16 @@
 import { TaskRepository } from '../../repositories/task-repository';
+import { Injectable, NotFoundException } from '@nestjs/common';
 
+@Injectable()
 export class DeleteTaskUseCase {
   constructor(readonly repository: TaskRepository) {}
 
-  execute(id: string) {
-    this.repository.delete(id);
+  async execute(id: string): Promise<void> {
+    const task = await this.repository.getById(id);
+    if (!task) {
+      throw new NotFoundException('Task not found');
+    }
+
+    await this.repository.delete(id);
   }
 }
