+++ src/domain/application/use-cases/user/delete-user.ts
@@ -1,9 +1,18 @@
-import { UserProps, UserRepository } from '../../repositories/user-repository';
+import { UserRepository } from '../../repositories/user-repository';
+import { Injectable, NotFoundException } from '@nestjs/common';
 
+@Injectable()
 export class DeleteUserUseCase {
   constructor(readonly repository: UserRepository) {}
 
-  execute(id: string) {
-    this.repository.delete(id);
+  async execute(id: string): Promise<void> {
+    // Verificando se o usuário existe antes de deletar
+    const user = await this.repository.getById(id);
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
+
+    // Excluindo o usuário
+    await this.repository.delete(id);
   }
 }
